Struct nunny::NonEmpty

pub struct NonEmpty<T: ?Sized> { /* private fields */ }

A wrapper struct around non-empty slices/arrays/vectors.

You may wish to use the following type aliases instead:

Slice.

Vec.

Array.

See also crate documentation

Implementations

impl<const N: usize, T> NonEmpty<[T; N]>

Array methods

pub const fn new_ref(src: &[T; N]) -> Option<&Self>

Returns a NonEmpty array.

Returns None if src is empty.

pub fn new_mut(src: &mut [T; N]) -> Option<&mut Self>

Returns a NonEmpty array.

Returns None if src is empty.

pub fn new(src: [T; N]) -> Option<Self>

Returns a NonEmpty array.

Returns None if src is empty.

pub const unsafe fn new_ref_unchecked(src: &[T; N]) -> &Self

Create a NonEmpty array.

Safety

src must not be empty

pub unsafe fn new_mut_unchecked(src: &mut [T; N]) -> &mut Self

Create a NonEmpty array.

Safety

src must not be empty

pub const unsafe fn new_unchecked(src: [T; N]) -> Self

Create a NonEmpty array.

Safety

src must not be empty

pub fn each_ref(&self) -> Array<&T, N>

Borrows each element and returns a NonEmpty array of references with the same size as self.

pub fn each_mut(&mut self) -> Array<&mut T, N>

Borrows each element mutably and returns a NonEmpty array of mutable references with the same size as self.

pub fn each_map<F, U>(self, f: F) -> Array<U, N>
where F: FnMut(T) -> U,

Returns a NonEmpty array of the same size as self, with function f applied to each element in order.

pub const fn as_slice_ne(&self) -> &Slice<T>

Returns a NonEmpty slice.

pub fn as_mut_slice_ne(&mut self) -> &mut Slice<T>

Returns a NonEmpty slice.

pub const fn as_array(&self) -> &[T; N]

Returns a primitive array.

pub fn as_mut_array(&mut self) -> &mut [T; N]

Returns a primitive array.

pub fn into_array(self) -> [T; N]

Returns a primitive array.

impl<T, const N: usize> NonEmpty<[T; N]>

Known non-empty iterator for Array.

pub fn into_iter_ne(self) -> NonEmpty<IntoIter<T, N>>

impl<T> NonEmpty<[T; 1]>

Special case for Arrays of length one

pub const fn of(item: T) -> Self

Create a NonEmpty array of a single element

pub fn of_mut(item: &mut T) -> &mut Self

Create a NonEmpty array of a single mutable reference

pub const fn of_ref(item: &T) -> &Self

Create a NonEmpty array of a single reference

impl<I> NonEmpty<I>
where I: Iterator,

Methods on Iterators with a non-empty invariant.

See Self::relax to access the normal iterator inside.

pub fn first(self) -> I::Item

NonEmpty version of Iterator::next.

let v = vec![1, 2, 3];
let _: Option<&u8> = v.iter().next();
    // ^ normally you have to handle the empty case
let _: &u8 = v.iter_ne().first();
    // ^ but we know there is at least one element

pub fn last(self) -> I::Item

NonEmpty version of Iterator::last.

let v = vec![1, 2, 3];
let _: Option<&u8> = v.iter().last();
    // ^ normally you have to handle the empty case
let _: &u8 = v.iter_ne().last();
    // ^ but we know there is at least one element

pub fn map<B, F>(self, f: F) -> NonEmpty<Map<I, F>>
where F: FnMut(I::Item) -> B,

NonEmpty version of Iterator::map.

let iter = slice![1, 2, 3].iter_ne();
assert_eq!(
    iter.map(|it| *it * 2).last(),
                        // ^ the invariant is maintained
                        //   so we _know_ there's a last element
    6
);

pub fn chain<U>( self, other: U ) -> NonEmpty<Chain<I, <U as IntoIterator>::IntoIter>>
where U: IntoIterator<Item = I::Item>,

NonEmpty version of Iterator::chain.

let iter = slice![1, 2].iter_ne();
assert_eq!(
    iter.chain(&[3]).last(),
                  // ^ the invariant is maintained
                  //   so we _know_ there's a last element
    &3
);

pub fn enumerate(self) -> NonEmpty<Enumerate<I>>

NonEmpty version of Iterator::enumerate.

let iter = slice!['a', 'b'].iter_ne();
assert_eq!(
    iter.enumerate().last(),
                 // ^ the invariant is maintained
                 //   so we _know_ there's a last element
    (1, &'b')
);

pub fn peekable(self) -> NonEmpty<Peekable<I>>

NonEmpty version of Iterator::peekable, allowing you to use Self::peek and Self::peek_mut

let mut peek_me = vec!['a', 'b'].into_iter_ne().peekable();
assert_eq!(
    *peek_me.peek(),
    'a'
);
*peek_me.peek_mut() = 'b';
assert_eq!(
    peek_me.collect_vec(),
    ['b', 'b']
);

pub fn take(self, n: NonZeroUsize) -> NonEmpty<Take<I>>

NonEmpty version of Iterator::take.

Note that n cannot be zero, to maintain the NonEmpty invariant.

let iter = slice!['a', 'b'].iter_ne();
assert_eq!(
    iter.take(nonzero!(1)).last(),
           // ^ compile time checked
    &'a'
)

pub fn flatten<II, T>(self) -> NonEmpty<FlatMap<I, II, fn(_: I::Item) -> II>>
where I: Iterator<Item = NonEmpty<II>>, II: IntoIterator<Item = T>,

NonEmpty version of Iterator::flatten.

Note that the inner items must also be NonEmpty, to maintain the invariant.

use nunny::{vec};
let nested = vec![vec![1], vec![2, 3]];
assert_eq!(
    nested.into_iter_ne().flatten().collect_vec(),
    [1, 2, 3],
);

pub fn fuse(self) -> NonEmpty<Fuse<I>>

NonEmpty version of Iterator::fuse.

pub fn inspect<F>(self, f: F) -> NonEmpty<Inspect<I, F>>
where F: FnMut(&I::Item),

NonEmpty version of Iterator::inspect.

pub fn reduce<F>(self, f: F) -> I::Item
where F: FnMut(I::Item, I::Item) -> I::Item,

NonEmpty version of Iterator::reduce.

let v = vec![1, 2, 3];
let _: Option<&u8> = v.iter().reduce(min);
    // ^ normally you have to handle the empty case
let _: &u8 = v.iter_ne().reduce(min);
    // ^ but we know there is at least one element

pub fn max(self) -> I::Item
where I::Item: Ord,

NonEmpty version of Iterator::max.

let v = vec![1, 2, 3];
let _: Option<&u8> = v.iter().max();
    // ^ normally you have to handle the empty case
let _: &u8 = v.iter_ne().max();
    // ^ but we know there is at least one element

pub fn min(self) -> I::Item
where I::Item: Ord,

NonEmpty version of Iterator::min.

let v = vec![1, 2, 3];
let _: Option<&u8> = v.iter().min();
    // ^ normally you have to handle the empty case
let _: &u8 = v.iter_ne().min();
    // ^ but we know there is at least one element

pub fn max_by_key<B, F>(self, f: F) -> I::Item
where B: Ord, F: FnMut(&I::Item) -> B,

NonEmpty version of Iterator::max_by_key.

pub fn max_by<F>(self, compare: F) -> I::Item
where F: FnMut(&I::Item, &I::Item) -> Ordering,

NonEmpty version of Iterator::max_by.

pub fn min_by_key<B, F>(self, f: F) -> I::Item
where B: Ord, F: FnMut(&I::Item) -> B,

NonEmpty version of Iterator::min_by_key.

pub fn min_by<F>(self, compare: F) -> I::Item
where F: FnMut(&I::Item, &I::Item) -> Ordering,

NonEmpty version of Iterator::min_by.

pub fn rev(self) -> NonEmpty<Rev<I>>
where I: DoubleEndedIterator,

NonEmpty version of Iterator::rev.

pub fn unzip_vec<A, B>(self) -> (Vec<A>, Vec<B>)
where I: Iterator<Item = (A, B)>,

NonEmpty version of Iterator::unzip.

pub fn copied<'a, T>(self) -> NonEmpty<Copied<I>>
where T: 'a + Copy, I: Iterator<Item = &'a T>,

NonEmpty version of Iterator::copied.

pub fn cloned<'a, T>(self) -> NonEmpty<Cloned<I>>
where T: 'a + Clone, I: Iterator<Item = &'a T>,

NonEmpty version of Iterator::cloned.

pub fn cycle(self) -> NonEmpty<Cycle<I>>
where I: Clone,

NonEmpty version of Iterator::cycle.

pub fn relax(self) -> I

Remove the NonEmpty wrapper, allowing you to access normal iterator methods like Iterator::filter.

pub fn collect_vec(self) -> Vec<I::Item>

Collect this iterator into a NonEmpty<Vec>.

pub fn try_collect_vec<T, E>(self) -> Result<Vec<T>, E>
where I: Iterator<Item = Result<T, E>>,

Collect Ok items into a NonEmpty<Vec>, short-circuiting on Err.

impl<I> NonEmpty<Peekable<I>>
where I: Iterator,

pub fn peek(&mut self) -> &I::Item

Peek this NonEmpty iterator, without advancing it.

See Self::peekable.

pub fn peek_mut(&mut self) -> &mut I::Item

Peek and modify this NonEmpty iterator, without advancing it.

See Self::peekable.

impl<T> NonEmpty<[T]>

Slice methods

pub const fn new(src: &[T]) -> Option<&Self>

Create a new NonEmpty slice

Returns None if src is empty.

pub fn new_mut(src: &mut [T]) -> Option<&mut Self>

Create a new NonEmpty slice

Returns None if src is empty.

pub const unsafe fn new_unchecked(src: &[T]) -> &Self

Create a new NonEmpty slice

Safety

src must not be empty

pub unsafe fn new_mut_unchecked(src: &mut [T]) -> &mut Self

Create a new NonEmpty slice

Safety

src must not be empty

pub const fn of(item: &T) -> &Self

Create a NonEmpty slice of a single element

pub fn of_mut(item: &mut T) -> &mut Self

Create a NonEmpty slice of a single element

pub const fn as_slice(&self) -> &[T]

Returns a primitive slice.

pub fn as_mut_slice(&mut self) -> &mut [T]

Returns a primitive slice.

pub const fn len_ne(&self) -> NonZeroUsize

Returns the known non-zero length.

pub const fn first(&self) -> &T

Returns the first element, guaranteed.

pub const fn split_first(&self) -> (&T, &[T])

Returns the first element, guaranteed, and the rest of the elements.

pub const fn split_last(&self) -> (&T, &[T])

Returns the last element, guaranteed, and the rest of the elements.

pub const fn last(&self) -> &T

Returns the last element, guaranteed.

pub fn first_mut(&mut self) -> &mut T

Returns the first element, guaranteed.

pub fn split_first_mut(&mut self) -> (&mut T, &mut [T])

Returns the first element, guaranteed, and the rest of the elements.

pub fn split_last_mut(&mut self) -> (&mut T, &mut [T])

Returns the last element, guaranteed, and the rest of the elements.

pub fn last_mut(&mut self) -> &mut T

Returns the last element, guaranteed.

impl<T> NonEmpty<[T]>

Known non-empty iterators for Slice.

pub fn iter_ne(&self) -> NonEmpty<Iter<'_, T>>

pub fn iter_mut_ne(&mut self) -> NonEmpty<IterMut<'_, T>>

pub fn into_iter_ne(self: Box<Self>) -> NonEmpty<IntoIter<T>>

impl<T> NonEmpty<Vec<T>>

Vec methods

pub fn new_ref(src: &Vec<T>) -> Option<&Self>

Create a new NonEmpty heap-allocated vec

Returns None if src is empty.

pub fn new_mut(src: &mut Vec<T>) -> Option<&mut Self>

Create a new NonEmpty heap-allocated vec

Returns None if src is empty.

pub unsafe fn new_unchecked(src: Vec<T>) -> Self

Create a new NonEmpty heap-allocated vec

Safety

src must not be empty

pub unsafe fn new_ref_unchecked(src: &Vec<T>) -> &Self

Create a new NonEmpty heap-allocated vec

Safety

src must not be empty

pub unsafe fn new_mut_unchecked(src: &mut Vec<T>) -> &mut Self

Create a new NonEmpty heap-allocated vec

Safety

src must not be empty

pub fn new(src: Vec<T>) -> Result<Self, Vec<T>>

Create a new NonEmpty heap-allocated vec, returning the original allocation if it was empty.

pub fn of(item: T) -> Self

Create a NonEmpty heap-allocated vec, of a single element.

pub fn of_with_capacity(item: T, capacity: usize) -> Self

Create a NonEmpty heap-allocated vec, of a single element, with capacity for capacity elements without (re)-allocating.

pub fn of_extending<A>(first: T, rest: impl IntoIterator<Item = A>) -> Self
where Self: Extend<A>,

Creating a NonEmpty heap-allocated vec where the first element is known.

This is an convenience method, equivalent to

let mut it = nunny::vec!["first"];
it.extend(["this", "is", "the", "rest"]);

pub fn filled(value: T, len: NonZeroUsize) -> Self
where T: Clone,

Create a NonEmpty heap-allocated vec with len items, filled with Clones of the given value.

See also Self::filled_with.

pub fn filled_with<F>(f: F, len: NonZeroUsize) -> Self
where F: FnMut() -> T,

Create a NonEmpty heap-allocated vec with len items, filled with values returned from repeating the closure f.

See also Self::filled.

pub fn as_vec(&self) -> &Vec<T>

Returns a std::vec::Vec.

pub unsafe fn as_mut_vec(&mut self) -> &mut Vec<T>

Returns a std::vec::Vec.

Safety

returned vec must not be emptied through this reference

pub fn into_vec(self) -> Vec<T>

Returns a std::vec::Vec.

pub fn as_slice_ne(&self) -> &Slice<T>

Returns a NonEmpty slice.

pub fn as_mut_slice_ne(&mut self) -> &mut Slice<T>

Returns a NonEmpty slice.

pub fn capacity(&self) -> NonZeroUsize

Returns the known non-zero length.

pub fn reserve(&mut self, additional: usize)

See reserve.

pub fn reserve_exact(&mut self, additional: usize)

See reserve_exact.

pub fn try_reserve(&mut self, additional: usize) -> Result<(), TryReserveError>

See try_reserve.

pub fn try_reserve_exact( &mut self, additional: usize ) -> Result<(), TryReserveError>

See try_reserve_exact.

pub fn shrink_to_fit(&mut self)

See shrink_to_fit.

pub fn shrink_to(&mut self, min_capacity: usize)

See shrink_to.

pub fn into_boxed_slice(self) -> Box<Slice<T>>

Return a NonEmpty boxed slice.

pub fn truncate(&mut self, len: NonZeroUsize)

Shortens the vector to a guaranteed-nonzero length

See truncate.

pub unsafe fn set_len(&mut self, new_len: NonZeroUsize)

Safety

See set_len.

pub fn insert(&mut self, index: usize, element: T)

See insert.

pub fn dedup_by_key<F, K>(&mut self, key: F)
where F: FnMut(&mut T) -> K, K: PartialEq,

See dedup_by_key.

pub fn dedup_by<F>(&mut self, same_bucket: F)
where F: FnMut(&mut T, &mut T) -> bool,

See dedup_by.

pub fn push(&mut self, value: T)

See push.

pub fn append(&mut self, other: &mut Vec<T>)

See append.

pub fn resize_with<F>(&mut self, new_len: NonZeroUsize, f: F)
where F: FnMut() -> T,

See resize_with.

pub fn leak<'a>(self) -> &'a mut Slice<T>

Returns a NonEmpty slice.

See leak.

pub fn spare_capacity_mut(&mut self) -> &mut [MaybeUninit<T>]

See spare_capacity_mut.

impl<T> NonEmpty<Vec<T>>

Known non-empty iterator for Vec.

pub fn into_iter_ne(self) -> NonEmpty<IntoIter<T>>