Type Alias nunny::Array

pub type Array<T, const N: usize> = NonEmpty<[T; N]>;

A non-empty array of known size.

Aliased Type

struct Array<T, const N: usize> { /* private fields */ }

Implementations

impl<const N: usize, T> Array<T, N>

Array methods

pub const fn new_ref(src: &[T; N]) -> Option<&Self>

Returns a NonEmpty array.

Returns None if src is empty.

pub fn new_mut(src: &mut [T; N]) -> Option<&mut Self>

Returns a NonEmpty array.

Returns None if src is empty.

pub fn new(src: [T; N]) -> Option<Self>

Returns a NonEmpty array.

Returns None if src is empty.

pub const unsafe fn new_ref_unchecked(src: &[T; N]) -> &Self

Create a NonEmpty array.

Safety

src must not be empty

pub unsafe fn new_mut_unchecked(src: &mut [T; N]) -> &mut Self

Create a NonEmpty array.

Safety

src must not be empty

pub const unsafe fn new_unchecked(src: [T; N]) -> Self

Create a NonEmpty array.

Safety

src must not be empty

pub fn each_ref(&self) -> Array<&T, N>

Borrows each element and returns a NonEmpty array of references with the same size as self.

pub fn each_mut(&mut self) -> Array<&mut T, N>

Borrows each element mutably and returns a NonEmpty array of mutable references with the same size as self.

pub fn each_map<F, U>(self, f: F) -> Array<U, N>
where F: FnMut(T) -> U,

Returns a NonEmpty array of the same size as self, with function f applied to each element in order.

pub const fn as_slice_ne(&self) -> &Slice<T>

Returns a NonEmpty slice.

pub fn as_mut_slice_ne(&mut self) -> &mut Slice<T>

Returns a NonEmpty slice.

pub const fn as_array(&self) -> &[T; N]

Returns a primitive array.

pub fn as_mut_array(&mut self) -> &mut [T; N]

Returns a primitive array.

pub fn into_array(self) -> [T; N]

Returns a primitive array.

impl<T, const N: usize> Array<T, N>

Known non-empty iterator for Array.

pub fn into_iter_ne(self) -> NonEmpty<IntoIter<T, N>>

impl<T> Array<T, 1>

Special case for Arrays of length one

pub const fn of(item: T) -> Self

Create a NonEmpty array of a single element

pub fn of_mut(item: &mut T) -> &mut Self

Create a NonEmpty array of a single mutable reference

pub const fn of_ref(item: &T) -> &Self

Create a NonEmpty array of a single reference