Type Alias nunny::Vec

pub type Vec<T> = NonEmpty<Vec<T>>;

A non-empty, heap allocated Vec.

Aliased Type

struct Vec<T> { /* private fields */ }

Implementations

impl<T> Vec<T>

Vec methods

pub fn new_ref(src: &Vec<T>) -> Option<&Self>

Create a new NonEmpty heap-allocated vec

Returns None if src is empty.

pub fn new_mut(src: &mut Vec<T>) -> Option<&mut Self>

Create a new NonEmpty heap-allocated vec

Returns None if src is empty.

pub unsafe fn new_unchecked(src: Vec<T>) -> Self

Create a new NonEmpty heap-allocated vec

Safety

src must not be empty

pub unsafe fn new_ref_unchecked(src: &Vec<T>) -> &Self

Create a new NonEmpty heap-allocated vec

Safety

src must not be empty

pub unsafe fn new_mut_unchecked(src: &mut Vec<T>) -> &mut Self

Create a new NonEmpty heap-allocated vec

Safety

src must not be empty

pub fn new(src: Vec<T>) -> Result<Self, Vec<T>>

Create a new NonEmpty heap-allocated vec, returning the original allocation if it was empty.

pub fn of(item: T) -> Self

Create a NonEmpty heap-allocated vec, of a single element.

pub fn of_with_capacity(item: T, capacity: usize) -> Self

Create a NonEmpty heap-allocated vec, of a single element, with capacity for capacity elements without (re)-allocating.

pub fn of_extending<A>(first: T, rest: impl IntoIterator<Item = A>) -> Self
where Self: Extend<A>,

Creating a NonEmpty heap-allocated vec where the first element is known.

This is an convenience method, equivalent to

let mut it = nunny::vec!["first"];
it.extend(["this", "is", "the", "rest"]);

pub fn filled(value: T, len: NonZeroUsize) -> Self
where T: Clone,

Create a NonEmpty heap-allocated vec with len items, filled with Clones of the given value.

See also Self::filled_with.

pub fn filled_with<F>(f: F, len: NonZeroUsize) -> Self
where F: FnMut() -> T,

Create a NonEmpty heap-allocated vec with len items, filled with values returned from repeating the closure f.

See also Self::filled.

pub fn as_vec(&self) -> &Vec<T>

Returns a std::vec::Vec.

pub unsafe fn as_mut_vec(&mut self) -> &mut Vec<T>

Returns a std::vec::Vec.

Safety

returned vec must not be emptied through this reference

pub fn into_vec(self) -> Vec<T>

Returns a std::vec::Vec.

pub fn as_slice_ne(&self) -> &Slice<T>

Returns a NonEmpty slice.

pub fn as_mut_slice_ne(&mut self) -> &mut Slice<T>

Returns a NonEmpty slice.

pub fn capacity(&self) -> NonZeroUsize

Returns the known non-zero length.

pub fn reserve(&mut self, additional: usize)

See reserve.

pub fn reserve_exact(&mut self, additional: usize)

See reserve_exact.

pub fn try_reserve(&mut self, additional: usize) -> Result<(), TryReserveError>

See try_reserve.

pub fn try_reserve_exact( &mut self, additional: usize ) -> Result<(), TryReserveError>

See try_reserve_exact.

pub fn shrink_to_fit(&mut self)

See shrink_to_fit.

pub fn shrink_to(&mut self, min_capacity: usize)

See shrink_to.

pub fn into_boxed_slice(self) -> Box<Slice<T>>

Return a NonEmpty boxed slice.

pub fn truncate(&mut self, len: NonZeroUsize)

Shortens the vector to a guaranteed-nonzero length

See truncate.

pub unsafe fn set_len(&mut self, new_len: NonZeroUsize)

Safety

See set_len.

pub fn insert(&mut self, index: usize, element: T)

See insert.

pub fn dedup_by_key<F, K>(&mut self, key: F)
where F: FnMut(&mut T) -> K, K: PartialEq,

See dedup_by_key.

pub fn dedup_by<F>(&mut self, same_bucket: F)
where F: FnMut(&mut T, &mut T) -> bool,

See dedup_by.

pub fn push(&mut self, value: T)

See push.

pub fn append(&mut self, other: &mut Vec<T>)

See append.

pub fn resize_with<F>(&mut self, new_len: NonZeroUsize, f: F)
where F: FnMut() -> T,

See resize_with.

pub fn leak<'a>(self) -> &'a mut Slice<T>

Returns a NonEmpty slice.

See leak.

pub fn spare_capacity_mut(&mut self) -> &mut [MaybeUninit<T>]

See spare_capacity_mut.

impl<T> Vec<T>

Known non-empty iterator for Vec.

pub fn into_iter_ne(self) -> NonEmpty<IntoIter<T>>